+++ gatsby/util/append-trailing-slash.js
@@ -1,0 +1,7 @@
+const appendTrailingSlash = (path) => {
+  return path.endsWith('/') ? path : `${path}/`;
+};
+
+module.exports = {
+  appendTrailingSlash,
+};
